Code improvements

+++ social-network-api/src/main/java/com/jonatas/socialnetworkapi/services/WorkerService.java
@@ -3,6 +3,7 @@ package com.jonatas.socialnetworkapi.services;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
@@ -10,54 +11,74 @@ import com.jonatas.socialnetworkapi.dto.WorkerDTO;
 import com.jonatas.socialnetworkapi.entities.Entity;
 import com.jonatas.socialnetworkapi.entities.User;
 import com.jonatas.socialnetworkapi.entities.Worker;
-import com.jonatas.socialnetworkapi.repositories.EntityRepository;
-import com.jonatas.socialnetworkapi.repositories.UserRepository;
 import com.jonatas.socialnetworkapi.repositories.WorkerRepository;
 
 @Service
 public class WorkerService {
 
+	//repositories
+	
 	@Autowired
 	private WorkerRepository workerRepository;
 	
-	@Autowired
-	private UserRepository userRepository;
+	//services
 	
 	@Autowired
-	private EntityRepository entityRepository;
+	private UserService userService;
+	
+	@Autowired
+	private EntityService entityService;
+	
+	//methods
 	
 	public ResponseEntity<List<Worker>> findAll(){
 		List<Worker> list = workerRepository.findAll();
 		return ResponseEntity.ok(list);
 	}
 	
+	public ResponseEntity<Worker> findById(String id){
+		try {
+			Worker worker = workerRepository.findById(id).get();
+			return ResponseEntity.ok().body(worker);
+		}catch (RuntimeException e) {
+			return ResponseEntity.notFound().build();
+		}
+	}
+	
 	public ResponseEntity<Worker> create(WorkerDTO workerDTO) {
 		try {
-			User user = userRepository.findById(workerDTO.getUser()).get();
-			Entity entity = entityRepository.findById(workerDTO.getEntity()).get();
-			Worker worker = new Worker(null, user, entity, workerDTO.getRole());
+			User user = userService.findById(workerDTO.getUser()).getBody();
+			Entity entity = entityService.findById(workerDTO.getEntity()).getBody();
+			Worker worker = new Worker(null, workerDTO.getRole(), user, entity);
+			if(!user.isChecked()) {
+				return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+			}
 			workerRepository.save(worker);
 			user.getWorkers().add(worker);
-			userRepository.save(user);
+			userService.save(user);
 			entity.getWorkers().add(worker);
-			entityRepository.save(entity);
+			entityService.save(entity);
 			return ResponseEntity.created(null).body(worker);
+			
 		}catch(RuntimeException e) {
 			return ResponseEntity.badRequest().build();
 		}
 		
 	}
 	
-	public ResponseEntity<Void> delete(String id){
+	public ResponseEntity<Void> delete(String idWorker, String idUser){
 		try {
-			Worker worker = workerRepository.findById(id).get();
+			Worker worker = workerRepository.findById(idWorker).get();
 			User user = worker.getUser();
+			if(idUser.hashCode() != user.getId().hashCode()) {
+				return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+			}
 			user.getWorkers().remove(worker);
 			Entity entity = worker.getEntity();
 			entity.getWorkers().remove(worker);
 			workerRepository.delete(worker);
-			userRepository.save(user);
-			entityRepository.save(entity);
+			userService.save(user);
+			entityService.save(entity);
 			return ResponseEntity.ok().build();
 		}catch(RuntimeException e) {
 			return ResponseEntity.notFound().build();
